factor out query param checking, begin to add query param to embeds
