+++ web/embed-flutter.html
@@ -139,6 +139,7 @@ BSD-style license that can be found in the LICENSE file. -->
             <div id="test-view" class="flex-auto d-flex flex-row no-overflow" spellcheck="false" hidden>
                 <div id="test-editor" class="flex-auto d-flex flex-column" spellcheck="false"></div>
             </div>
+            <div id="feature-message" class="feature-message"></div>
         </div>
         <div class="editor-button-group">
             <button id="copy-code" title="Copy code" class="mdc-icon-button material-icons">file_copy</button>
